style(schedule): improve notification animation
This commit enhances the animation for the cache update notification on the schedule screen.

The previous `easeInOut` animation has been replaced with a more dynamic `spring` animation to provide a more fluid and natural user experience. The view's transition has also been made asymmetric for a smoother appearance and disappearance.